Move basic input components into separate directory
